Move LSP message impls into submodules
